requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
fastapi
uvicorn[standard]
starlette
aiofiles
annotated-types
gunicorn
python-multipart
jinja2==3.0.3
requests>=2.31.0
pydantic>=2.0.0
email-validator>=2.0.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Image Processing
Pillow>=10.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0